fix: verify the carried history chain before re-signing and pruning
The prune takes an array prefix, so it depended on the oldest-first ordering
being true, and that ordering was prose in the blueprint rather than a check.
A committed writer could publish a newest-first or gapped chain, pass the gate
- detached signatures authenticate links individually, never their sequence -
and have the next honest rotator drop the newest links and strand the ratchet.

A rotation now walks the carried chain before touching it. The re-sealer holds
one key into the chain, the previous epoch's seed, and each link yields the key
for the link before it, so a wrong order, a gap, or a foreign link all surface
as an open that fails. Ordering is proven rather than assumed, and the prune
demonstrably drops the oldest end. Nothing is re-signed sight-unseen.

The codec cannot make this check: a link's epoch lives in its untransmitted AAD
and inside its ciphertext, so crates/core sees an opaque sealed blob. A sweep
cannot either - it holds only the scope's current seed, never a key into the
carried chain - so it now neither walks nor prunes, and since it appends
nothing the set cannot grow there.

Also cast the two remaining ordinal-to-bytes fixtures to u64. usize is 4 bytes
on wasm32, so the kat_gen site would have emitted different vectors on a 32-bit
host - a determinism hole in a generator whose output is a frozen contract.
Vectors are byte-identical after the cast.

+++ crates/core/examples/kat_gen.rs
@@ -4918,7 +4918,7 @@ fn build_grant_section_reject() -> Vec<RejectVector> {
         (
             "history-links-past-the-bound",
             section_with_links(
-                (0..=seal::MAX_HISTORY_LINKS)
+                (0..=seal::MAX_HISTORY_LINKS as u64)
                     .map(|i| history_link_map(i.to_be_bytes().to_vec()))
                     .collect(),
             ),
